pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "surge-surrogate"
version = "0.1.0"
description = "Surrogate Unified Robust Generation Engine - A modular AI/ML framework for building fast, accurate, and uncertainty-aware surrogate models"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Álvaro Sánchez Villar", email = "your.email@example.com"}
]
maintainers = [
    {name = "Álvaro Sánchez Villar", email = "your.email@example.com"}
]
keywords = ["machine learning", "surrogate models", "scientific computing", "neural networks", "gaussian processes"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.5.0",
    "torch>=1.9.0",
    "joblib>=1.0.0",
    "pyyaml>=5.0",
    "optuna>=3.0.0",
]

[project.optional-dependencies]
pytorch = [
    "torch>=1.9.0",
    "torchvision",
    "torchaudio",
]
gpflow = [
    "tensorflow>=2.8.0,<2.16",
    "tensorflow-probability>=0.19.0",
    "gpflow>=2.5.0",
]
shap = [
    "shap>=0.43.0",
]
botorch = [
    "botorch>=0.8.0",
    "gpytorch>=1.9.0",
    "optuna-integration[botorch]>=3.5.0",
]
viz = [
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
]
onnx = [
    "onnxruntime>=1.14.0",
]
mlflow = [
    "mlflow>=2.0.0",
]
data = [
    "h5py>=3.0.0",
    "xarray>=2023.0.0",
    "netCDF4>=1.6.0",
]
all = [
    "torch>=1.9.0",
    "torchvision",
    "torchaudio",
    "tensorflow>=2.8.0,<2.16",
    "tensorflow-probability>=0.19.0",
    "gpflow>=2.5.0",
    "shap>=0.43.0",
    "botorch>=0.8.0",
    "gpytorch>=1.9.0",
    "optuna-integration[botorch]>=3.5.0",
    "scikit-optimize>=0.9.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "h5py>=3.0.0",
    "xarray>=2023.0.0",
    "netCDF4>=1.6.0",
    "onnxruntime>=1.14.0",
    "psutil",
    "tqdm",
]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "ruff",
    "isort",
    "mypy",
    "jupyter",
    "matplotlib",
    "seaborn",
    "pandas",
    "optuna",
    "scikit-optimize",
    "botorch",
    "psutil",
    "tqdm",
]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "numpydoc",
]

[project.scripts]
surge = "surge.cli:cli"

[project.urls]
Homepage = "https://github.com/your-username/SURGE"
Documentation = "https://surge-surrogate.readthedocs.io/"
Repository = "https://github.com/your-username/SURGE.git"
"Bug Tracker" = "https://github.com/your-username/SURGE/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["surge*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# Default: normal mode (shows test names), use -q for quiet or -v for verbose
addopts = "--tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*.py" = ["F401", "F811"]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["surge"]
omit = [
    "tests/*",
    "setup.py",
    "docs/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
